+++ cmd/ftl/cmd_schema_protobuf.go
@@ -1,0 +1,14 @@
+package main
+
+import (
+	"fmt"
+
+	"github.com/TBD54566975/ftl/backend/schema"
+)
+
+type schemaProtobufCmd struct{}
+
+func (c *schemaProtobufCmd) Run() error { //nolint:unparam
+	fmt.Println(schema.ProtobufSchema())
+	return nil
+}
